+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Sanitize `http.url` attribute. ([#1961](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1961))
+
 ## 1.0.0-rc3
 
 Released 2021-Mar-19